+++ agent/impl/src/main/java/com/walmartlabs/concord/agent/ServerConnector.java
@@ -28,7 +28,6 @@ import com.walmartlabs.concord.ApiClient;
 import com.walmartlabs.concord.client.CommandQueueApi;
 import com.walmartlabs.concord.client.ProcessApi;
 import com.walmartlabs.concord.client.ProcessQueueApi;
-import com.walmartlabs.concord.project.InternalConstants;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;